Se cambiaron ubicaciones de inicializaciones
