+++ SistemaRiego/src/Drivers/waterPump.c
@@ -1,0 +1,14 @@
+/*
+ * waterPump.c
+ *
+ *  Created on: 13 de dic. de 2016
+ *      Author: mmaccio
+ */
+
+#include "infotronic.h"
+
+void InitBomba()
+{
+	//conf_gpio(PORT_BOMB,PORT_BOMB,OUT);
+	//setPin(PORT_BOMB,PORT_BOMB,0);
+}
