+++ SistemaRiego/src/Headers/rtc.h
@@ -59,6 +59,8 @@
 
 #define PCRTC 9
 
+void init_rtc(uint8_t ss, uint8_t mm, uint8_t hh, uint8_t dd, uint8_t mo, uint16_t yy);
+
 void config_date_RTC(uint8_t ss, uint8_t mm, uint8_t hh, uint8_t dd, uint8_t mo, uint16_t yy);
 
 #endif /* HEADERS_RTC_H_ */
