+++ SistemaRiego/src/Headers/waterPump.h
@@ -1,0 +1,13 @@
+/*
+ * waterPump.h
+ *
+ *  Created on: 13 de dic. de 2016
+ *      Author: mmaccio
+ */
+
+#ifndef HEADERS_WATERPUMP_H_
+#define HEADERS_WATERPUMP_H_
+
+void InitBomba();
+
+#endif /* HEADERS_WATERPUMP_H_ */
